Feat: add computer face on nav hover
